fix inv_sbox bug
in the old method of decrypting the 0th element of the table was being
set seperately instead of the 99th one (0x63) as 0x63 is the affine
transformation of 0 and 0 does not have an inverse. this bug did not
make a change as inv_sbox was being initialized on the stack with 0's
and the 99th was being automatically set to 0. in the new method
inv_sbox is initialized on the heap and filled with random data
therefore the 99th is incorrect.